LayoutTests/fast/viewport/viewport-77-expected.txt

CONSOLE MESSAGE: line 3: Viewport width or height set to physical device width, try using "device-width" constant instead for future compatibility.
CONSOLE MESSAGE: line 3: Viewport argument value "yes;" for key "initial-scale" not recognized. Content ignored.
viewport size 1280x1408 scale 0.250000 with limits [0.250000, 5.000000]
